+++ src/client/arguments/mod.rs
@@ -1,22 +1,2 @@
-pub mod access;
-pub mod commit;
-pub mod create;
-pub mod fs_info;
-pub mod fs_stat;
-pub mod get_attr;
-pub mod link;
-pub mod lookup;
-pub mod mk_dir;
-pub mod mk_node;
 pub mod mount;
-pub mod path_conf;
-pub mod read;
-pub mod read_dir;
-pub mod read_dir_plus;
-pub mod read_link;
-pub mod remove;
-pub mod rename;
-pub mod rm_dir;
-pub mod set_attr;
-pub mod symlink;
-pub mod write;
+pub mod nfsv3;
