+++ src/client/arguments/nfsv3/create.rs
@@ -5,7 +5,7 @@ use crate::serializer::files::dir_op_arg;
 use crate::serializer::{array, u32};
 use crate::vfs::create::{Args, How, Verifier};
 
-use super::set_attr::serialize_new_attr;
+use crate::client::arguments::nfsv3::set_attr::serialize_new_attr;
 
 /// Serializes [`Verifier`].
 fn serialize_verifier(dest: &mut impl Write, verf: Verifier) -> Result<()> {
