+++ src/client/arguments/nfsv3/mk_dir.rs
@@ -3,7 +3,7 @@ use std::io::{Result, Write};
 use crate::serializer::files::dir_op_arg;
 use crate::vfs::mk_dir::Args;
 
-use super::set_attr::serialize_new_attr;
+use crate::client::arguments::nfsv3::set_attr::serialize_new_attr;
 
 /// Serializes the arguments [`Args`] for an NFSv3 `ACCESS` operation to the provided `Write` destination.
 pub fn mk_dir_args(dest: &mut impl Write, arg: Args) -> Result<()> {
