+++ src/client/arguments/nfsv3/symlink.rs
@@ -1,6 +1,6 @@
 use std::io::{Result, Write};
 
-use super::set_attr::serialize_new_attr;
+use crate::client::arguments::nfsv3::set_attr::serialize_new_attr;
 use crate::serializer::files::dir_op_arg;
 use crate::serializer::files::file_path;
 use crate::vfs::symlink::Args;
